refactor: Rename ingest-pipeline to pipeline and extract utility modules
- Rename ingest-pipeline.ts to pipeline.ts for brevity
- Rename createIngestPipeline to createPipeline
- Extract embedder logic to separate embedder.ts module
- Extract retry logic to separate retry.ts module
- Update exports in index.ts to use new module names
