+++ packages/rag/src/ingest/pipeline.test.ts
@@ -3,9 +3,9 @@ import type { ChunkStore } from "../chunk-store/types";
 import type { ChunkerFunction } from "../chunker/types";
 import type { DocumentLoader } from "../document-loader/types";
 import type { EmbedderFunction } from "../embedder/types";
-import { createIngestPipeline } from "./ingest-pipeline";
+import { createPipeline } from "./pipeline";
 
-describe("createIngestPipeline", () => {
+describe("createPipeline", () => {
 	let mockDocumentLoader: DocumentLoader<{ path: string }>;
 	let mockChunker: ChunkerFunction;
 	let mockEmbedder: EmbedderFunction;
@@ -14,9 +14,13 @@ describe("createIngestPipeline", () => {
 	beforeEach(() => {
 		mockDocumentLoader = {
 			// @ts-expect-error - this is a mock
-			*load() {
-				yield { content: "doc1", metadata: { path: "file1.txt" } };
-				yield { content: "doc2", metadata: { path: "file2.txt" } };
+			*loadMetadata() {
+				yield { path: "file1.txt" };
+				yield { path: "file2.txt" };
+			},
+			// @ts-expect-error - this is a mock
+			async loadDocument(metadata) {
+				return { content: `doc for ${metadata.path}`, metadata };
 			},
 		};
 
@@ -34,16 +38,16 @@ describe("createIngestPipeline", () => {
 	});
 
 	it("should process documents through the pipeline", async () => {
-		const ingest = createIngestPipeline({
+		const ingest = createPipeline({
 			documentLoader: mockDocumentLoader,
 			chunker: mockChunker,
 			embedder: mockEmbedder,
 			chunkStore: mockChunkStore,
-			documentKey: (doc) => doc.metadata.path,
+			documentKey: (metadata) => metadata.path,
 			metadataTransform: (metadata) => metadata,
 		});
 
-		const result = await ingest({});
+		const result = await ingest();
 
 		expect(result.totalDocuments).toBe(2);
 		expect(result.successfulDocuments).toBe(2);
@@ -62,18 +66,18 @@ describe("createIngestPipeline", () => {
 				.mockResolvedValueOnce(undefined),
 		};
 
-		const ingest = createIngestPipeline({
+		const ingest = createPipeline({
 			documentLoader: mockDocumentLoader,
 			chunker: mockChunker,
 			embedder: mockEmbedder,
 			chunkStore: failingChunkStore,
-			documentKey: (doc) => doc.metadata.path,
+			documentKey: (metadata) => metadata.path,
 			metadataTransform: (metadata) => metadata,
 			maxRetries: 2,
 			retryDelay: 10,
 		});
 
-		const result = await ingest({});
+		const result = await ingest();
 
 		expect(result.successfulDocuments).toBe(2);
 		expect(failingChunkStore.insert).toHaveBeenCalledTimes(3); // 1 fail + 1 success for first doc, 1 success for second doc
@@ -82,17 +86,17 @@ describe("createIngestPipeline", () => {
 	it("should call progress callback", async () => {
 		const onProgress = vi.fn();
 
-		const ingest = createIngestPipeline({
+		const ingest = createPipeline({
 			documentLoader: mockDocumentLoader,
 			chunker: mockChunker,
 			embedder: mockEmbedder,
 			chunkStore: mockChunkStore,
-			documentKey: (doc) => doc.metadata.path,
+			documentKey: (metadata) => metadata.path,
 			metadataTransform: (metadata) => metadata,
 			onProgress,
 		});
 
-		await ingest({});
+		await ingest();
 
 		expect(onProgress).toHaveBeenCalled();
 		const lastCall = onProgress.mock.calls[onProgress.mock.calls.length - 1][0];
@@ -100,17 +104,17 @@ describe("createIngestPipeline", () => {
 	});
 
 	it("should handle batch processing", async () => {
-		const ingest = createIngestPipeline({
+		const ingest = createPipeline({
 			documentLoader: mockDocumentLoader,
 			chunker: mockChunker,
 			embedder: mockEmbedder,
 			chunkStore: mockChunkStore,
-			documentKey: (doc) => doc.metadata.path,
+			documentKey: (metadata) => metadata.path,
 			metadataTransform: (metadata) => metadata,
 			maxBatchSize: 1,
 		});
 
-		await ingest({});
+		await ingest();
 
 		// With batch size 1 and 2 chunks per document, should call embedMany 4 times
 		expect(mockEmbedder.embedMany).toHaveBeenCalledTimes(4);
